requirements.txt
# Web Framework
fastapi
uvicorn[standard]
gunicorn

# AI
anthropic

# Database
supabase
# Research
tavily-python
yfinance
requests
# Auth
python-jose[cryptography]
passlib[bcrypt]

# Document Processing
PyPDF2
python-docx
PyMuPDF

# Utils
python-dotenv
pydantic
pydantic-settings
python-multipart

# Optional: Embeddings (for vector search)
sentence-transformers==2.2.2